+++ ChangeLog.md
@@ -15,6 +15,7 @@
   - Zugkontrolle-Parameter `S` muss zusätzlich `Display` implementieren.
   - Anpassen der `Nachricht`-Typen.
   - Umbenennen `SteuerungAktualisierenFehler::Deserialisieren` -> `Reservieren`.
+  - Umbenennen `Ergebnis::FehlerMitErsatzwert` -> `WertMitWarnungen`.
 - Dunkles Thema. Dazu wurden u.a. die `zeichnen` Methode des `Cache` angepasst.
 - Verwende PickList für Modus-Auswahl.
 - Verwende `rfd` zur Datei-Auswahl beim Speichern und Laden.